requirements.txt
flask
flask-Testing
flask-SQLAlchemy
flask-restful
flask-script
flask-marshmallow
marshmallow-sqlalchemy